+++ src/read.c
@@ -8,18 +8,19 @@
 
 #include <Rinternals.h>
 
-/* avoid protection issues with setAttrib where new symbols may trigger GC probelms */
+// avoid protection issues with setAttrib
+// where new symbols may trigger GC probelms
 static void setAttr(SEXP x, const char *name, SEXP val) {
-    PROTECT(val);
-    setAttrib(x, Rf_install(name), val);
-    UNPROTECT(1);
+  PROTECT(val);
+  setAttrib(x, Rf_install(name), val);
+  UNPROTECT(1);
 }
 
-/* add information attributes accorsing to the TIFF tags.
-   Only a somewhat random set (albeit mostly baseline) is supported */
+// Add information attributes according to the TIFF tags.
+// Only a somewhat random set (albeit mostly baseline) is supported
 static void TIFF_add_info(TIFF *tiff, SEXP res) {
-  uint32 i32;
-  uint16 i16;
+  uint32_t i32;
+  uint16_t i16;
   float f;
   char *c = 0;
 
@@ -58,57 +59,57 @@ static void TIFF_add_info(TIFF *tiff, SEXP res) {
 	    setAttr(res, "planar_config", mkString(uv));
 	  }
   }
-    if (TIFFGetField(tiff, TIFFTAG_COMPRESSION, &i16)) {  // working here
-	char uv[24];
-	const char *name = 0;
-	switch (i16) {
-	case 1: name = "none"; break;
-	case 2: name = "CCITT RLE"; break;
-	case 32773: name = "PackBits"; break;
-	case 3: name = "CCITT Group 3 fax"; break;
-	case 4: name = "CCITT Group 4 fax"; break;
-	case 5: name = "LZW"; break;
-	case 6: name = "old JPEG"; break;
-	case 7: name = "JPEG"; break;
-	case 8: name = "deflate"; break;
-	case 9: name = "JBIG b/w"; break;
-	case 10: name = "JBIG color"; break;
-	default:
-	    snprintf(uv, sizeof(uv), "unknown (%d)", i16);
-	    name = uv;
-	}
-	setAttr(res, "compression", mkString(name));
-    }
-    if (TIFFGetField(tiff, TIFFTAG_THRESHHOLDING, &i16))
-	setAttr(res, "threshholding", ScalarInteger(i16));
-    if (TIFFGetField(tiff, TIFFTAG_XRESOLUTION, &f))
-	setAttr(res, "x_resolution", ScalarReal(f));
-    if (TIFFGetField(tiff, TIFFTAG_YRESOLUTION, &f))
-	setAttr(res, "y_resolution", ScalarReal(f));
-    if (TIFFGetField(tiff, TIFFTAG_RESOLUTIONUNIT, &i16)) {
-	const char *name = "unknown";
-	switch (i16) {
-	case 1: name = "none"; break;
-	case 2: name = "inch"; break;
-	case 3: name = "cm"; break;
-	}
+  if (TIFFGetField(tiff, TIFFTAG_COMPRESSION, &i16)) {
+  	char uv[24];
+  	const char *name = 0;
+  	switch (i16) {
+    	case 1: name = "none"; break;
+    	case 2: name = "CCITT RLE"; break;
+    	case 32773: name = "PackBits"; break;
+    	case 3: name = "CCITT Group 3 fax"; break;
+    	case 4: name = "CCITT Group 4 fax"; break;
+    	case 5: name = "LZW"; break;
+    	case 6: name = "old JPEG"; break;
+    	case 7: name = "JPEG"; break;
+    	case 8: name = "deflate"; break;
+    	case 9: name = "JBIG b/w"; break;
+    	case 10: name = "JBIG color"; break;
+    	default:
+  	    snprintf(uv, sizeof(uv), "unknown (%d)", i16);
+  	    name = uv;
+	  }
+	  setAttr(res, "compression", mkString(name));
+  }
+  if (TIFFGetField(tiff, TIFFTAG_THRESHHOLDING, &i16))
+	  setAttr(res, "threshholding", ScalarInteger(i16));
+  if (TIFFGetField(tiff, TIFFTAG_XRESOLUTION, &f))
+	  setAttr(res, "x_resolution", ScalarReal(f));
+  if (TIFFGetField(tiff, TIFFTAG_YRESOLUTION, &f))
+	  setAttr(res, "y_resolution", ScalarReal(f));
+  if (TIFFGetField(tiff, TIFFTAG_RESOLUTIONUNIT, &i16)) {
+	  const char *name = "unknown";
+  	switch (i16) {
+  	  case 1: name = "none"; break;
+  	  case 2: name = "inch"; break;
+	    case 3: name = "cm"; break;
+	  }
 	setAttr(res, "resolution_unit", mkString(name));
-    }
-#ifdef TIFFTAG_INDEXED /* very recent in libtiff even though it's an old tag */
+  }
+  #ifdef TIFFTAG_INDEXED /* very recent in libtiff even though it's an old tag */
     if (TIFFGetField(tiff, TIFFTAG_INDEXED, &i16))
-	setAttr(res, "indexed", ScalarLogical(i16));
-#endif
-    if (TIFFGetField(tiff, TIFFTAG_ORIENTATION, &i16)) {
-	const char *name = "<invalid>";
-	switch (i16) {
-	case 1: name = "top_left"; break;
-	case 2: name = "top_right"; break;
-	case 3: name = "bottom_right"; break;
-	case 4: name = "bottom_left"; break;
-	case 5: name = "left_top"; break;
-	case 6: name = "right_top"; break;
-	case 7: name = "right_bottom"; break;
-	case 8: name = "left_bottom"; break;
+	    setAttr(res, "indexed", ScalarLogical(i16));
+  #endif
+  if (TIFFGetField(tiff, TIFFTAG_ORIENTATION, &i16)) {
+  	const char *name = "<invalid>";
+	  switch (i16) {
+  	case 1: name = "top_left"; break;
+	  case 2: name = "top_right"; break;
+  	case 3: name = "bottom_right"; break;
+	  case 4: name = "bottom_left"; break;
+  	case 5: name = "left_top"; break;
+	  case 6: name = "right_top"; break;
+  	case 7: name = "right_bottom"; break;
+	  case 8: name = "left_bottom"; break;
 	}
 	setAttr(res, "orientation", mkString(name));
     }
@@ -147,35 +148,45 @@ static void TIFF_add_info(TIFF *tiff, SEXP res) {
 }
 
 SEXP read_tif_c(SEXP sFn /*filename*/) {
+  uint64_t to_unprotect = 0;
   check_type_sizes();
-  SEXP res = R_NilValue, multi_res = R_NilValue, multi_tail = R_NilValue, dim;
+  SEXP res = PROTECT(R_NilValue), multi_res = PROTECT(R_NilValue);
+  SEXP multi_tail = PROTECT(R_NilValue), dim;
+  to_unprotect += 3;
   const char *fn;
   int n_img = 0;
   tiff_job_t rj;
   TIFF *tiff;
   FILE *f;
-	if (TYPEOF(sFn) != STRSXP || LENGTH(sFn) != 1) Rf_error("invalid filename");
+	if (TYPEOF(sFn) != STRSXP || LENGTH(sFn) != 1) {
+	  Rf_error("invalid filename");
+	  UNPROTECT(to_unprotect);
+	  return R_NilValue;
+	}
 	fn = CHAR(STRING_ELT(sFn, 0));
 	f = fopen(fn, "rb");
 	if (!f) {
 	  Rf_error("unable to open %s", fn);
+	  UNPROTECT(to_unprotect);
 	  return R_NilValue;
 	}
 	rj.f = f;
   tiff = TIFF_Open("rmc", &rj); /* no mmap, no chopping */
   if (!tiff) {
+    TIFFClose(tiff);
     Rf_error("Unable to open TIFF");
+    UNPROTECT(to_unprotect);
     return R_NilValue;
   }
 
   while (true) { /* loop over separate image in a directory if desired */
-  	uint32 imageWidth = 0, imageLength = 0, imageDepth;
-  	uint32 tileWidth, tileLength;
-  	uint32 x, y;
-  	uint16 config, bps = 8, spp = 1, sformat = 1, out_spp;
+  	uint32_t imageWidth = 0, imageLength = 0, imageDepth;
+  	uint32_t tileWidth, tileLength;
+  	uint32_t x, y;
+  	uint16_t config, bps = 8, spp = 1, sformat = 1, out_spp;
   	tdata_t buf;
   	double *real_arr;
-  	uint16 *colormap[3] = {0, 0, 0};
+  	uint16_t *colormap[3] = {0, 0, 0};
   	bool is_float = false;
 
   	TIFFGetField(tiff, TIFFTAG_IMAGEWIDTH, &imageWidth);
@@ -196,8 +207,11 @@ SEXP read_tif_c(SEXP sFn /*filename*/) {
   	  is_float = true;
   	}
   	if (spp == 1) { /* modify out_spp for colormaps */
-  	  if (colormap[2]) out_spp = 3;
-  	  else if (colormap[1]) out_spp = 2;
+  	  if (colormap[2]) {
+  	    out_spp = 3;
+  	  } else if (colormap[1]) {
+  	    out_spp = 2;
+  	  }
   	}
     #if TIFF_DEBUG
   	  Rprintf("image %d x %d x %d, tiles %d x %d, bps = %d, spp = %d (output %d), "
@@ -210,12 +224,13 @@ SEXP read_tif_c(SEXP sFn /*filename*/) {
       Rf_error("12-bit images are not supported. "
                "Try converting your image to 16-bit.");
       TIFFClose(tiff);
+      UNPROTECT(to_unprotect);
       return R_NilValue;
     }
   	if (bps != 8 && bps != 16 && bps != 32) {
   	    TIFFClose(tiff);
-  	    Rf_error("image has %d bits/sample which is unsupported in direct mode - "
-                  "use native=TRUE or convert=TRUE", bps);
+  	    Rf_error("image has %d bits/sample which is unsupported", bps);
+  	    UNPROTECT(to_unprotect);
   	    return R_NilValue;
   	}
 
@@ -225,6 +240,7 @@ SEXP read_tif_c(SEXP sFn /*filename*/) {
                     "the signed integer format.");
 
   	res = PROTECT(allocVector(REALSXP, imageWidth * imageLength * out_spp));
+  	to_unprotect++;
   	real_arr = REAL(res);
 
   	if (tileWidth == 0) {
@@ -238,7 +254,7 @@ SEXP read_tif_c(SEXP sFn /*filename*/) {
       #endif
     	for (strip = 0; strip < TIFFNumberOfStrips(tiff); strip++) {
     	  tsize_t n = TIFFReadEncodedStrip(tiff, strip, buf, (tsize_t) -1);
-    	  if (spp == 1) { /* config doesn't matter for spp == 1 */
+    	  if (spp == 1) { // config doesn't matter for spp == 1
     	    if (colormap[0]) {
     	  	  tsize_t i, step = bps / 8;
     			  for (i = 0; i < n; i += step) {
@@ -306,7 +322,7 @@ SEXP read_tif_c(SEXP sFn /*filename*/) {
       			  }
       			}
     		  }
-    		} else if (config == PLANARCONFIG_CONTIG) { /* interlaced */   //working here
+    		} else if (config == PLANARCONFIG_CONTIG) { // interlaced
     		  tsize_t i, j, step = spp * bps / 8;
     		  for (i = 0; i < n; i += step) {
     			  const uint8_t *v = (const uint8_t*) buf + i;
@@ -348,7 +364,7 @@ SEXP read_tif_c(SEXP sFn /*filename*/) {
     			      (uint16_t) ((const uint16_t*)v)[0];
     			  }	else if (bps == 32 && !is_float) {
     			    real_arr[plane_offset + imageLength * x + y] =
-    			      (uint32_t) ((const uint32*)v)[0];
+    			      (uint32_t) ((const uint32_t*)v)[0];
     			  }	else if (bps == 32 && is_float) {
     			    real_arr[plane_offset + imageLength * x + y] =
     			      (double) ((const float*)v)[0];
@@ -366,19 +382,15 @@ SEXP read_tif_c(SEXP sFn /*filename*/) {
     		}
     	}
   	} else {
+  	  TIFFClose(tiff);
   	  Rf_error("tile-based images are not supported");
-  	  buf = _TIFFmalloc(TIFFTileSize(tiff));
-
-  	  for (y = 0; y < imageLength; y += tileLength) {
-  		  for (x = 0; x < imageWidth; x += tileWidth) {
-  		    TIFFReadTile(tiff, buf, x, y, 0 /*depth*/, 0 /*plane*/);
-  		  }
-  	  }
+  	  UNPROTECT(to_unprotect);
+  	  return R_NilValue;
   	}
 
   	_TIFFfree(buf);
 
-  	dim = allocVector(INTSXP, (out_spp > 1) ? 3 : 2);
+  	dim = PROTECT(allocVector(INTSXP, (out_spp > 1) ? 3 : 2));
   	INTEGER(dim)[0] = imageLength;
   	INTEGER(dim)[1] = imageWidth;
   	if (out_spp > 1) INTEGER(dim)[2] = out_spp;
@@ -386,25 +398,28 @@ SEXP read_tif_c(SEXP sFn /*filename*/) {
     TIFF_add_info(tiff, res);
   	UNPROTECT(1);
   	n_img++;
-  	if (multi_res == R_NilValue) {
-  	  multi_tail = multi_res = CONS(res, R_NilValue);
-  	  PROTECT(multi_res);
+  	if (multi_res == R_NilValue) {  // first image in stack
+  	  multi_res = multi_tail = PROTECT(CONS(res, R_NilValue));
+  	  to_unprotect ++;
   	} else {
-  	  SEXP q = CONS(res, R_NilValue);
-  	  SETCDR(multi_tail, q);
+  	  SEXP q = PROTECT(CONS(res, R_NilValue));
+  	  SETCDR(multi_tail, q);  // q is now protected as part of multi_tail
   	  multi_tail = q;
+  	  UNPROTECT(2);  // removing explit protection of q and protection of res
+  	  to_unprotect--;
   	}
   	if (!TIFFReadDirectory(tiff)) break;
   }
   TIFFClose(tiff);
   /* convert LISTSXP into VECSXP */
-  PROTECT(res = allocVector(VECSXP, n_img));
+  res = PROTECT(allocVector(VECSXP, n_img));
+  to_unprotect++;
 	int i = 0;
 	while (multi_res != R_NilValue) {
 	  SET_VECTOR_ELT(res, i, CAR(multi_res));
 	  i++;
 	  multi_res = CDR(multi_res);
 	}
-  UNPROTECT(2);
+  UNPROTECT(to_unprotect);
   return res;
 }