+++ NEWS.md
@@ -1,3 +1,14 @@
+### 0.3.0
+
+#### MINOR IMPROVEMENTS
+* Improve README and vignette with more tangible and fun example.
+
+#### BUG FIXES
+* Fix windows `libtiff` issues (thanks to Jeroen Ooms).
+* Found some ImageJ-written TIFFs that weren't being read correctly and fixed that.
+* Fix `protection stack overflow` error for TIFFs with many images.
+
+
 ## 0.2.0
 * First CRAN release.
 
@@ -7,4 +18,5 @@
 
 
 # `ijtiff` 0.1.0
-* First github release
+
+* First github release.